+++ modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/util/BaseQueryContext.java
@@ -221,7 +221,7 @@ public final class BaseQueryContext extends AbstractQueryContext {
         return rexBuilder;
     }
 
-    public long schemaVersion() {
+    public int schemaVersion() {
         return Objects.requireNonNull(schema().unwrap(IgniteSchema.class)).schemaVersion();
     }
